# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Refresh access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session terminated"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/restore": {
      "post": {
        "operationId": "AuthController_restore",
        "summary": "Restore deleted account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "summary": "Verify email with token from registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "summary": "Request password reset email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If email exists, reset link will be sent"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password with token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password has been reset"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Intake API",
    "description": "API contract for Intake",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd!",
            "minLength": 8,
            "maxLength": 80
          },
          "sex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01"
          },
          "height": {
            "type": "number",
            "example": 175,
            "minimum": 120,
            "maximum": 230
          },
          "weight": {
            "type": "number",
            "example": 70,
            "minimum": 30,
            "maximum": 300
          },
          "goal": {
            "type": "string",
            "enum": [
              "lose",
              "maintain",
              "gain"
            ],
            "example": "lose"
          },
          "activityLevel": {
            "type": "number",
            "enum": [
              1.2,
              1.375,
              1.55,
              1.725,
              1.9
            ],
            "example": 1.55
          },
          "targetWeight": {
            "type": "number",
            "minimum": 30,
            "maximum": 300
          },
          "goalDelta": {
            "type": "number",
            "minimum": -600,
            "maximum": 600
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "sex",
          "dateOfBirth",
          "height",
          "weight",
          "goal",
          "activityLevel"
        ]
      },
      "AuthTokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "MetabolismDto": {
        "type": "object",
        "properties": {
          "bmr": {
            "type": "number",
            "example": 1500
          },
          "recommendedCalories": {
            "type": "number",
            "example": 2000
          }
        },
        "required": [
          "bmr",
          "recommendedCalories"
        ]
      },
      "UserSettingsDto": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "example": "light"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "pl",
              "uk"
            ],
            "example": "en"
          },
          "sound": {
            "type": "boolean",
            "example": true
          },
          "volume": {
            "type": "number",
            "example": 25
          }
        },
        "required": [
          "theme",
          "language",
          "sound",
          "volume"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0c1b2a3e4f56789012345"
          },
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "sex": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1990-01-01"
          },
          "age": {
            "type": "number",
            "example": 30
          },
          "height": {
            "type": "number",
            "example": 175
          },
          "weight": {
            "type": "number",
            "example": 70
          },
          "targetWeight": {
            "type": "number",
            "example": 60
          },
          "goal": {
            "type": "string",
            "enum": [
              "lose",
              "maintain",
              "gain"
            ],
            "example": "lose"
          },
          "goalDelta": {
            "type": "number",
            "example": 5
          },
          "activityLevel": {
            "type": "number",
            "enum": [
              1.2,
              1.375,
              1.55,
              1.725,
              1.9
            ],
            "example": 1.55
          },
          "metabolism": {
            "$ref": "#/components/schemas/MetabolismDto"
          },
          "settings": {
            "$ref": "#/components/schemas/UserSettingsDto"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "sex",
          "dateOfBirth",
          "age",
          "height",
          "weight",
          "targetWeight",
          "goal",
          "goalDelta",
          "activityLevel",
          "metabolism",
          "settings"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensResponseDto"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "tokens",
          "user"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "EMAIL_ALREADY_EXISTS",
              "INVALID_CREDENTIALS",
              "INVALID_CURRENT_PASSWORD",
              "NEW_PASSWORD_MUST_BE_DIFFERENT",
              "UNAUTHORIZED",
              "INVALID_SESSION",
              "VALIDATION_ERROR",
              "USER_NOT_FOUND",
              "ACCOUNT_DELETED",
              "INVALID_RESET_TOKEN",
              "INVALID_VERIFICATION_TOKEN",
              "FOOD_NOT_FOUND",
              "FOOD_BAD_REQUEST",
              "AI_PARSE_FAILED",
              "DAY_NOT_FOUND",
              "SERVER_ERROR"
            ],
            "example": "VALIDATION_ERROR",
            "description": "Machine-readable error code"
          }
        },
        "required": [
          "code"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd!",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Email verification token from registration email"
          }
        },
        "required": [
          "token"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@email.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token from email"
          },
          "newPassword": {
            "type": "string",
            "example": "P@ssw0rd!",
            "minLength": 8,
            "maxLength": 80
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
    }
  }
}